+++ c/verifythis/elimination_max.c
@@ -1,4 +1,4 @@
-extern void *calloc(unsigned long nmemb, unsigned long size);
+extern void *calloc(unsigned int nmemb, unsigned int size);
 extern void __VERIFIER_error(void) __attribute__ ((__noreturn__));
 extern void __VERIFIER_assume(int);
 void __VERIFIER_assert(int cond) {
